updated file structure and library
